+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add result timestamp (when check is started)
 * Rename http(s) attributes: statusCodes -> statusCode, texts -> text
 * All checker attributes now support <attribute> and <attribute>-or
+* Add health#clearCache
 
 ### 0.2.1
 * Add file checker